requirements.txt
Flask
dash
dash-core-components
dash-html-components
psycopg2-binary